extract commissioning.py from matter_testing.py
